+++ src/Aks/Aks.Test/ScenarioTests/KubernetesTests.cs
@@ -149,5 +149,12 @@ namespace Commands.Aks.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-EnableFIPS");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestAutoScalerProfile()
+        {
+            TestRunner.RunTestScript("Test-AutoScalerProfile");
+        }
     }
 }